+++ test-playwright/VvFormFieldsGroup.v3.vue
@@ -5,13 +5,15 @@ import type { Ref } from 'vue'
 import { useForm } from '../src'
 import NameSurname from './components/NameSurname.vue'
 
-defineEmits(['submit', 'invalid', 'valid', 'reset'])
-
+const props = defineProps<{
+    initialData?: z.infer<typeof zodSchema>
+}>()
+const emit = defineEmits(['submit', 'invalid', 'valid', 'reset', 'mounted'])
 const zodSchema = z.object({
     firstname: z.string().min(3),
     lastname: z.string().min(3),
-})
-
+},
+)
 const { VvForm, VvFormFieldsGroup } = useForm(zodSchema, {
     lazyLoad: true,
 })
@@ -23,18 +25,17 @@ const model: Ref<z.infer<typeof zodSchema>> = ref({
 
 const formEl = ref<InstanceType<typeof VvForm>>()
 onMounted(() => {
-    if (formEl.value) {
-        formEl.value.submit()
+    if (!props.initialData) {
+        return
     }
+    model.value = props.initialData
+    emit('mounted', model.value)
 })
 </script>
 
 <template>
     <VvForm
-        ref="formEl"
-        v-model="model"
-        @submit="$emit('submit', $event)"
-        @invalid="$emit('invalid')"
+        ref="formEl" v-model="model" @submit="$emit('submit', $event)" @invalid="$emit('invalid')"
         @valid="$emit('valid')"
     >
         <VvFormFieldsGroup