+++ CHANGELOG.md
@@ -8,6 +8,10 @@ All notable changes to this project will be documented in this file.
 
 - Support for Zod 4 from `zod@3.25.x` and `zod@4.x.x`.
 
+### Fixed
+
+- `VvFormFieldGroup` initial value;
+
 ## [1.0.0] - 2024-07-21
 
 ### Added